requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
numpy==1.22.2
pandas==1.4.1
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
